// src/features/cart/types/Cart.ts
import type { Product } from '../../../types/product';

export type CartEntry = {
  name?: string;
  price?: number;
  product: Product;
  quantity?: number;
  total?: number;
  _id?: string;
};

export interface Cart {
  createdAt?: Date;
  items: CartEntry[];
  totalItems?: number;
  totalPrice?: number;
  updatedAt?: Date;
  user?: string;
  _id?: string;
}